backend/requirements.txt
# Smart Tank Backend - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# MQTT Client
paho-mqtt==1.6.1

# Data Processing
pandas==2.2.2
numpy==1.25.2
scikit-learn==1.3.2

# Time Series
influxdb-client==1.38.0

# Notifications
requests==2.31.0
python-telegram-bot==20.7

# AI
google-generativeai==0.8.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0